Change inventory implementation
The child nodes of Inventory are now considered to be the slots of the
Inventory rather than maintaining a seperate 'slots' member variable.

By doing this, InventorySlot no longer needs a reference to its inventory
or to keep a record of it's index. It can simply use get_parent() and
get_index() respectively.